# speed up package version check in `.rlang_cli_compat()`