@lionel- 2023-10-07
Thanks! The other thing you could do is cache the information in a top-level object namespaced with the compat prefix.